Implemented basic Google authentication
