+++ src/App.tsx
@@ -1,45 +1,45 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import logo from './logo.svg';
 import './App.css';
 import { initializeApp } from "firebase/app";
+import { auth } from './frontend/services/firebase';
+import { onAuthStateChanged } from 'firebase/auth';
+import Home from './frontend/components/Home';
+import SignIn from './frontend/components/SignIn'
 import { getAnalytics, logEvent } from "firebase/analytics";
+
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
 // Your web app's Firebase configuration
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: "AIzaSyA_YpeR-nCdjPFCbQREO3Bg9Dp4BIWUPpk",
-  authDomain: "oututoringdatabase.firebaseapp.com",
-  projectId: "oututoringdatabase",
-  storageBucket: "oututoringdatabase.appspot.com",
-  messagingSenderId: "763676444526",
-  appId: "1:763676444526:web:b0eac5aa6892e9a2ac237f",
-  measurementId: "G-CC2H30MC7N"
-};
+// const firebaseConfig = {
+//   apiKey: "AIzaSyA_YpeR-nCdjPFCbQREO3Bg9Dp4BIWUPpk",
+//   authDomain: "oututoringdatabase.firebaseapp.com",
+//   projectId: "oututoringdatabase",
+//   storageBucket: "oututoringdatabase.appspot.com",
+//   messagingSenderId: "763676444526",
+//   appId: "1:763676444526:web:b0eac5aa6892e9a2ac237f",
+//   measurementId: "G-CC2H30MC7N"
+// };
 
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-logEvent(analytics, 'notification_received');
+// // Initialize Firebase
+// const app = initializeApp(firebaseConfig);
+// const analytics = getAnalytics(app);
+// logEvent(analytics, 'notification_received');
 
 function App() {
+  const [user, setUser] = useState<any>();
+
+  useEffect(() => {
+    onAuthStateChanged(auth, user => {
+      setUser(user);
+    })
+  }, []);
+
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      {user ? <Home user={user}/> : <SignIn />}
     </div>
   );
 }
